style: adjust loader code
