+++ packages/runtime/loader/src/utils.ts
@@ -1,4 +1,4 @@
-import { error, parseContentType } from '@garfish/utils';
+import { error, isObject, isPrimitive, parseContentType } from '@garfish/utils';
 import { Loader, LoadedPluginArgs } from './index';
 
 export async function request(url: string, config: RequestInit) {
@@ -25,4 +25,28 @@ export function mergeConfig(loader: Loader, url: string) {
   const extra = loader.requestConfig;
   const config = typeof extra === 'function' ? extra(url) : extra;
   return { mode: 'cors', ...config } as RequestInit;
+}
+
+export function calculateObjectSize(obj: any) {
+  let size = 0;
+  const valueSet = new WeakSet();
+  const add = (val: any) => {
+    if (isPrimitive(val) || typeof val === 'function') {
+      size += new Blob([val]).size;
+    } else if (isObject(val)) {
+      if (!valueSet.has(val)) {
+        valueSet.add(val);
+        for (const key in val) add(val[key]);
+      }
+    } else if (Array.isArray(val)) {
+      if (!valueSet.has(val)) {
+        valueSet.add(val);
+        val.forEach(add);
+      }
+    } else {
+      size += new Blob([val]).size;
+    }
+  };
+  add(obj);
+  return size;
 }
